feat/modify response of modules and lessons crud

+++ src/modules/modules.service.ts
@@ -2,13 +2,15 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateModuleDto } from './dto/create-module.dto';
 import { UpdateModuleDto } from './dto/update-module.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { GlobalException } from 'src/exceptions/global-exception';
 
 @Injectable()
 export class ModulesService {
   constructor(private prisma: PrismaService) {}
 
   async createModule(courseId: string, createModuleDto: CreateModuleDto) {
-    const { title } = createModuleDto;
+    try {
+      const { title } = createModuleDto;
     const course = await this.prisma.course.findUnique({
       where: { id: courseId },
       include: { Module: true },
@@ -23,32 +25,51 @@ export class ModulesService {
       },
     });
     return module;
+    } catch (error) {
+      throw new GlobalException("Failed to create module", error.message);
+    }
   }
 
   async getModuleById(id: string) {
-    const module = await this.prisma.module.findUnique({
-      where: { id },
-      include: { Lesson: true },
-    });
-    if (!module) {
-      throw new NotFoundException(`Module with id ${id} not found`);
+    try {
+      const module = await this.prisma.module.findUnique({
+        where: { id },
+        include: { Lesson: true },
+      });
+      if (!module) {
+        throw new NotFoundException(`Module with id ${id} not found`);
+      }
+      return module;
+    } catch (error) {
+      throw new GlobalException("Failed to get module", error.message);
     }
-    return module;
   }
 
   async updateModule(id: string, updateModuleDto: UpdateModuleDto) {
-    const { title } = updateModuleDto;
-    const module = await this.getModuleById(id);
-    return this.prisma.module.update({
-      where: { id },
-      data: { title },
-    });
+    try {
+      const { title } = updateModuleDto;
+      const module = await this.getModuleById(id);
+      if (!module) throw new NotFoundException(`Module with id ${id} not found`);
+      const updatedModule = await this.prisma.module.update({
+        where: { id },
+        data: { title },
+      });
+      return updatedModule;
+    } catch (error) {
+      throw new GlobalException("Failed to update module", error.message);
+    }
   }
 
   async deleteModule(id: string) {
-    const module = await this.getModuleById(id);
-    return this.prisma.module.delete({
-      where: { id },
-    });
+    try {
+      const module = await this.getModuleById(id);
+      if (!module) throw new NotFoundException(`Module with id ${id} not found`);
+      const deletedModule = await this.prisma.module.delete({
+        where: { id },
+      });
+      return deletedModule;
+    } catch (error) {
+      throw new GlobalException("Failed to delete module", error.message);
+    }
   }
 }
